Split canonical port protocols
Move canonical repository and unit-of-work protocols into focused modules so each file stays below the lint line ceiling. Keep `episodic.canonical.ports` as a compatibility re-export while updating internal imports to the new protocol modules.
