+++ episodic/api/runtime.py
@@ -18,7 +18,7 @@ if typ.TYPE_CHECKING:
 
     from falcon import asgi
 
-    from episodic.canonical.ports import CanonicalUnitOfWork
+    from episodic.canonical.unit_of_work_protocols import CanonicalUnitOfWork
 
     from .types import UowFactory
