+++ episodic/api/types.py
@@ -18,7 +18,7 @@ import collections.abc as cabc
 import typing as typ
 
 if typ.TYPE_CHECKING:
-    from episodic.canonical.ports import CanonicalUnitOfWork
+    from episodic.canonical.unit_of_work_protocols import CanonicalUnitOfWork
 
 type UowFactory = cabc.Callable[[], CanonicalUnitOfWork]
 type JsonPayload = dict[str, object]
