+++ episodic/architecture/policy.py
@@ -64,7 +64,11 @@ def _domain_ports_group() -> ModuleGroup:
             "episodic.canonical.constraints",
             "episodic.canonical.ingestion",
             "episodic.canonical.ingestion_ports",
+            "episodic.canonical.entity_protocols",
+            "episodic.canonical.history_protocols",
             "episodic.canonical.ports",
+            "episodic.canonical.reference_protocols",
+            "episodic.canonical.unit_of_work_protocols",
             "episodic.llm.ports",
         ),
         allowed_groups=frozenset({"domain_ports"}),
