+++ episodic/canonical/profile_templates/helpers.py
@@ -31,7 +31,7 @@ from .types import (
 
 if typ.TYPE_CHECKING:
     from episodic.canonical.domain import EpisodeTemplate, JsonMapping, SeriesProfile
-    from episodic.canonical.ports import CanonicalUnitOfWork
+    from episodic.canonical.unit_of_work_protocols import CanonicalUnitOfWork
 
 
 async def _get_latest_revision[HistoryEntryT: _RevisionedEntry](
