+++ episodic/canonical/profile_templates/services/_typed.py
@@ -43,7 +43,7 @@ from ._generic import (
 )
 
 if typ.TYPE_CHECKING:
-    from episodic.canonical.ports import CanonicalUnitOfWork
+    from episodic.canonical.unit_of_work_protocols import CanonicalUnitOfWork
 
 
 get_series_profile = partial(get_entity_with_revision, kind="series_profile")
