+++ episodic/canonical/reference_documents/documents.py
@@ -24,7 +24,7 @@ from .types import (
 )
 
 if typ.TYPE_CHECKING:
-    from episodic.canonical.ports import CanonicalUnitOfWork
+    from episodic.canonical.unit_of_work_protocols import CanonicalUnitOfWork
 
 
 async def create_reference_document(
