+++ episodic/canonical/reference_documents/helpers.py
@@ -20,7 +20,7 @@ from .types import (
 )
 
 if typ.TYPE_CHECKING:
-    from episodic.canonical.ports import CanonicalUnitOfWork
+    from episodic.canonical.unit_of_work_protocols import CanonicalUnitOfWork
 
 
 _MAX_PAGE_LIMIT = 100
