+++ tests/conftest.py
@@ -16,7 +16,7 @@ if typ.TYPE_CHECKING:
     import sqlalchemy as sa
 
     from episodic.canonical.domain import EpisodeTemplate
-    from episodic.canonical.ports import CanonicalUnitOfWork
+    from episodic.canonical.unit_of_work_protocols import CanonicalUnitOfWork
 pytest_plugins: list[str] = [
     "tests.fixtures.database",
     "tests.fixtures.llm",
