+++ tests/test_async_task_factory.py
@@ -42,7 +42,7 @@ from episodic.canonical.ingestion_service import (
 )
 
 if typ.TYPE_CHECKING:
-    from episodic.canonical.ports import CanonicalUnitOfWork
+    from episodic.canonical.unit_of_work_protocols import CanonicalUnitOfWork
 
 
 class _TaskConstructorKwargs(typ.TypedDict):
